requirements.txt
black==22.1.0
click==8.0.4
colorama==0.4.4
mypy-extensions==0.4.3
numpy==1.22.2
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
